+++ assets/src/components/apps/misc.tsx
@@ -5,6 +5,7 @@ import {
   IconFrame,
 } from '@pluralsh/design-system'
 import { Flex, P } from 'honorable'
+import { Readiness } from 'utils/status'
 
 export const hasIcons = ({ spec: { descriptor } }) => descriptor?.icons?.length > 0
 
@@ -16,6 +17,14 @@ export const getIcon = (app, dark) => {
   if (dark && descriptor.icons.length > 1) return descriptor.icons[1]
 
   return descriptor.icons[0]
+}
+
+export function appState({ status: { conditions } }) {
+  const ready = conditions.find(({ type }) => type === 'Ready')
+  const error = conditions.find(({ type }) => type === 'Error')
+  const readiness = error.status === 'True' ? Readiness.Failed : (ready.status === 'True' ? Readiness.Ready : Readiness.InProgress)
+
+  return { ready, error, readiness }
 }
 
 export function ListItemBorder({ borderColor }: { borderColor: string }) {
