--- assets/src/components/utils/Loading.jsx
@@ -1,15 +1,0 @@
-import { Loading as LoadingInner } from 'forge-core'
-import { Box } from 'grommet'
-
-export function Loading({ background }) {
-  return (
-    <Box
-      fill
-      background={background}
-      justify="center"
-      align="center"
-    >
-      <LoadingInner />
-    </Box>
-  )
-}
