--- assets/src/components/utils/icons.jsx
@@ -1,27 +1,0 @@
-import { Blank } from 'grommet-icons'
-
-export function Send(props) {
-  return (
-    <Blank {...props}>
-      <svg
-        version="1.1"
-        id="Layer_1"
-        x="0px"
-        y="0px"
-        width="24px"
-        height="24px"
-        viewBox="12 12 24 24"
-        enableBackground="new 12 12 24 24"
-      >
-        <path
-          fill="#32785C"
-          d="M14.51,33l20.99-9l-20.99-9l-0.01,7l15,2l-15,2L14.51,33z"
-        />
-        <path
-          fill="none"
-          d="M0,0h48v48H0V0z"
-        />
-      </svg>
-    </Blank>
-  )
-}
